Publish package to GitHub Packages without specifying registry
This commit updates the release.yml file to remove the explicit registry URL when publishing a package to GitHub Packages. Instead of using `npm.pkg.github.com/wizarrrr` as the registry, the default registry will be used. This change simplifies the publishing process and makes it consistent with other workflows.